+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## August 31, 2020
+- Modern JavaScript From The Beginning (Udemy)
+  - Section 5: Object Oriented JavaScript - ES5 & ES2015
+    - Constructors & the 'this' Keyword
+- An example with constructors and the this keyword was added.
+
 ## August 27, 2020
 - Modern JavaScript From The Beginning (Udemy)
   - Section 4: DOM Projects